Reverting data loader changes
