+++ test/dataset/test_variable_length.py
@@ -16,6 +16,7 @@ from functools import partial
 # Third-party imports
 import mxnet as mx
 import numpy as np
+import pandas as pd
 import pytest
 
 # First-party imports
@@ -25,8 +26,7 @@ from gluonts.dataset.loader import (
     TrainDataLoader,
     InferenceDataLoader,
 )
-from gluonts.dataset.parallelized_loader import batchify, stack
-from gluonts.testutil.dummy_datasets import get_dataset
+from gluonts.dataset.parallelized_loader import batchify, stack, _pad_arrays
 from gluonts.transform import (
     ContinuousTimeInstanceSplitter,
     ContinuousTimeUniformSampler,
@@ -35,6 +35,37 @@ from gluonts.transform import (
 
 @pytest.fixture
 def pp_dataset():
+    def get_dataset():
+
+        data_entry_list = [
+            {
+                "target": np.c_[
+                    np.array([0.2, 0.7, 0.2, 0.5, 0.3, 0.3, 0.2, 0.1]),
+                    np.array([0, 1, 2, 0, 1, 2, 2, 2]),
+                ].T,
+                "start": pd.Timestamp("2011-01-01 00:00:00", freq="H"),
+                "end": pd.Timestamp("2011-01-01 03:00:00", freq="H"),
+            },
+            {
+                "target": np.c_[
+                    np.array([0.2, 0.1, 0.2, 0.5, 0.4]),
+                    np.array([0, 1, 2, 1, 1]),
+                ].T,
+                "start": pd.Timestamp("2011-01-01 00:00:00", freq="H"),
+                "end": pd.Timestamp("2011-01-01 03:00:00", freq="H"),
+            },
+            {
+                "target": np.c_[
+                    np.array([0.2, 0.7, 0.2, 0.5, 0.1, 0.2, 0.1]),
+                    np.array([0, 1, 2, 0, 1, 0, 2]),
+                ].T,
+                "start": pd.Timestamp("2011-01-01 00:00:00", freq="H"),
+                "end": pd.Timestamp("2011-01-01 03:00:00", freq="H"),
+            },
+        ]
+
+        return ListDataset(data_entry_list, freq="H", one_dim_target=False,)
+
     return get_dataset
 
 
@@ -232,3 +263,21 @@ def test_variable_length_stack_zerosize(
     assert stacked.shape[0] == 5
     assert stacked.shape[1] == 1
     assert stacked.shape[2] == 2
+
+
+@pytest.mark.parametrize(
+    "array_type, multi_processing, axis",
+    itertools.product(["np", "mx"], [True, False], [0, 1]),
+)
+def test_pad_arrays_axis(pp_dataset, array_type, multi_processing, axis: int):
+    arrays = [
+        d["target"] if array_type == "np" else mx.nd.array(d["target"])
+        for d in list(iter(pp_dataset()))
+    ]
+    if axis == 0:
+        arrays = [x.T for x in arrays]
+
+    padded_arrays = _pad_arrays(arrays, axis)
+
+    assert all(a.shape[axis] == 8 for a in padded_arrays)
+    assert all(a.shape[1 - axis] == 2 for a in padded_arrays)
